backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-socketio
requests
beautifulsoup4
python-dotenv
google-generativeai

langchain>=0.2
langchain-core>=0.2
langchain-text-splitters>=0.2
langchain-community>=0.2
google-generativeai>=0.7
pandas>=2.0
requests>=2.31
